+++ src/coremods/core_whowas.cpp
@@ -443,9 +443,9 @@ class ModuleWhoWas : public Module, public Stats::EventListener
 	void ReadConfig(ConfigStatus& status) CXX11_OVERRIDE
 	{
 		ConfigTag* tag = ServerInstance->Config->ConfValue("whowas");
-		unsigned int NewGroupSize = tag->getUInt("groupsize", 10, 0, 10000);
-		unsigned int NewMaxGroups = tag->getUInt("maxgroups", 10240, 0, 1000000);
-		unsigned int NewMaxKeep = tag->getDuration("maxkeep", 3600, 3600);
+		unsigned int NewGroupSize = static_cast<unsigned int>(tag->getUInt("groupsize", 10, 0, 10000));
+		unsigned int NewMaxGroups = static_cast<unsigned int>(tag->getUInt("maxgroups", 10240, 0, 1000000));
+		unsigned int NewMaxKeep = static_cast<unsigned int>(tag->getDuration("maxkeep", 3600, 3600));
 
 		cmd.manager.UpdateConfig(NewGroupSize, NewMaxGroups, NewMaxKeep);
 	}
